% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robseq.R
\name{Robseq}
\alias{Robseq}
\title{Robseq Function}
\usage{
Robseq(
  features,
  metadata,
  norm.method = "tmm",
  expVar = "Exposure",
  coVars = NULL,
  parallel = FALSE,
  ncores = 1
)
}
\arguments{
\item{features}{Gene Expressions}

\item{metadata}{Native Tissue or Tumor}

\item{norm.method}{Normalizing method}

\item{expVar}{'Exposure' by default}

\item{coVars}{'NULL' by default}

\item{parallel}{'FALSE' by default}

\item{ncores}{'1' by default}
}
\value{
output
}
\description{
Robseq Function
}
