% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_Robseq.R
\name{Robseq}
\alias{Robseq}
\title{Robseq}
\usage{
Robseq(
  features,
  metadata,
  norm.method = "TMM",
  expVar = "Exposure",
  coVars = NULL,
  parallel = FALSE,
  ncores = 1
)
}
\arguments{
\item{features}{A dataframe with gene expression counts in the row and samples in the column.}

\item{metadata}{A dataframe with information on the subjects such as disease status, gender and etc.}

\item{norm.method}{The normalization method to be used. The user can choose from 5 different methods such as TMM, RLE, CPM, Upper quartile and Qauntile. ‘TMM’ by default.}

\item{expVar}{The name of the variable on which the differential expression will be evaluated. If the user provides no name then the metadata should have a column named as exposure which should have information on things such as disease status, treatment conditions or etc. ‘Exposure’ by default.}

\item{coVars}{The names of the covariates/confounders that needs to adjusted for in the differential expression analysis. 'NULL' by default}

\item{parallel}{If true, the analysis will be performed on multiple cores with faster runtimes. 'FALSE' by default}

\item{ncores}{The number of cores on which the analysis will be serially performed. The user needs to specify this only when parallel = TRUE. '1' by default}
}


\details{
Details Goes Here
}

\description{
\if{html}{\figure{RobseqLogo.png}{options: style='float: right' alt='Figure: RobseqLogo.png' width='120'}}

A Robust Statistical Model for Differential Gene Expression Analysis in RNA-Seq Studies
}
\value{
Values goes here
}

\author{
Suvo Chatterjee, Vrushab Hanumesh, Siddhant Sunil Meshram 

}
\references{
Reference goes here
}
\examples{
\dontrun{Robseq(features, metadata)}

}
